.mov of blockchain UI
Excuse the speed of the video... Had to go fast to make .mov under 25MB.
The video displays the streamlit user interface of the PyChain powered blockchain. In the side bar is the difficulty slider; which will increase the difficulty of the hash function, making it slower, and harder to hack. As well as a drop down menu to view a specific transaction.